Extract validation methods to a utilities class
